Move `ExtractIf` and version parsing to `utils` module
